+++ CHANGELOG.md
@@ -57,6 +57,9 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-build: Fixed an issue with rain not appearing in some maps
 - invader-build: Fixed an issue with detail objects not appearing in some maps
 - invader-build: Uncompressed bitmaps marked as compressed now error
+- invader-build: Fixed bend bounds defaulting to 0-1 rather than being based on
+  natural pitch (you will get a pedantic warning if it does this and your bend
+  bounds are not 0)
 - invader-edit-qt: Fixed crashing if a tag directory contained files that could
   not be accessed in the filesystem
 - invader-edit-qt: Fixed 8-bit color being displayed in the wrong order